Problem: publishing `emerald-cli` via `cargo install` not possbile
Problem: in order to allow distribution of an executable `emerald-cli`
via `cargo install` a separate crate needs to be created focusing only
on CLI functionality (with a dependency on emerald-core).

Secondly the core emerald library needs to expose the FFI in a shared
object and rust crate pub functions for crates.io. This allows users to
import only the core functionality into their apps without having to
bother with the cli functionality.

Solution: split out emerald into two crates `emerald-core` and
`emerald-cli`

Note the CLI functionaly isn't completed this is just to get this
architectural change done so that development can proceed without
throwing mutex locks over the codebase.