Problem: rustfmt on nix is a different version to CI
Problem: seems there isn't an easy way to set the version of
cargo fmt to be 0.8.4 on nix which is 0.7.1, despite trying a
number of times to manually override it in ~/.cargo/bin directory.

Solution: reset the history prior to running rust fmt on the code
base. Hopefully this issue won't persist, unless my editor made
some annoying changes. If that's the case then I'll have to either
manually change each offending line of code so that it satisfies CI or
find a way to get cargo fmt on nix to behave properly.