Problem: rustfmt fails on 1 line
Root Problem: enforcing rustfmt at CI is overly pedantic and
messes with development workflow, especially when developers
have different environments. No solution to this other than
unnecessary yak shaving or convincing others that running
rustfmt at CI is a bad idea. A better idea is to run
rustfmt once in a while as a tradeoff. Better yet don't run
rustfmt as it makes unnecessary changes to code which can conflic
with other programmers work. So all the amount of yak shaving
could be removed entirely instead of wasting hours trying to
appease the CI gods.

Temporary solution: fix the single line of code so that it
doesn't offend and cross my fingers it'll pass.